# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticate with email and password, returns JWT tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_auth_dto.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_auth_dto.LoginOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Exchange a valid refresh token for a new token pair",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh tokens",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_auth_dto.RefreshInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_auth_dto.RefreshOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Change password for an authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Change user password",
                "parameters": [
                    {
                        "description": "Current and new password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.ChangePasswordInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ServiceName": []
                    },
                    {
                        "ServiceKey": []
                    }
                ],
                "description": "Set initial password for a user that does not have one yet. Protected by Service Key (not JWT) to avoid deadlock: user needs a password to login, but needs to login to get JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Set user password",
                "parameters": [
                    {
                        "description": "User ID, password and confirmation",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.SetPasswordInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_DenysonJ_financial-wallet_internal_usecases_auth_dto.LoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "Email do usuário",
                    "type": "string"
                },
                "password": {
                    "description": "Senha em texto plano",
                    "type": "string"
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_auth_dto.LoginOutput": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "JWT access token (curta duração)",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "JWT refresh token (longa duração)",
                    "type": "string"
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_auth_dto.RefreshInput": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "description": "Refresh token atual",
                    "type": "string"
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_auth_dto.RefreshOutput": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "Novo JWT access token",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "Novo JWT refresh token",
                    "type": "string"
                }
            }
        },
...
        "github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.ChangePasswordInput": {
            "type": "object",
            "required": [
                "current_password",
                "new_password",
                "new_password_confirmation"
            ],
            "properties": {
                "current_password": {
                    "description": "Senha atual",
                    "type": "string"
                },
                "new_password": {
                    "description": "Nova senha",
                    "type": "string"
                },
                "new_password_confirmation": {
                    "description": "Confirmação da nova senha",
                    "type": "string"
                }
            }
        },
...
        "github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.SetPasswordInput": {
            "type": "object",
            "required": [
                "password",
                "password_confirmation",
                "user_id"
            ],
            "properties": {
                "password": {
                    "description": "Senha em texto plano",
                    "type": "string"
                },
                "password_confirmation": {
                    "description": "Confirmação da senha",
                    "type": "string"
                },
                "user_id": {
                    "description": "ID do usuário",
                    "type": "string"
                }
            }
        },
...
        "internal_infrastructure_web_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "example": "error description"
                        }
                    }
                }
            }
        }
    }
}